Add different spawn for each player
